migrate the configuration to be core capable
