+++ martin-core/src/resources/fonts/mod.rs
@@ -33,8 +33,6 @@ use pbf_font_tools::{Fontstack, Glyphs, render_sdf_glyph};
 use regex::Regex;
 use serde::{Deserialize, Serialize};
 
-use crate::config::OptOneMany;
-
 /// Maximum Unicode codepoint supported (U+FFFF - Basic Multilingual Plane).
 const MAX_UNICODE_CP: usize = 0xFFFF;
 /// Size of each Unicode codepoint range (256 characters).
@@ -114,7 +112,7 @@ pub struct CatalogFontEntry {
 }
 
 /// Thread-safe font manager for discovery, cataloging, and serving fonts as Protocol Buffers.
-#[derive(Debug, Clone, Default)]
+#[derive(Debug, Clone)]
 pub struct FontSources {
     /// Map of font name to font source data.
     fonts: DashMap<String, FontSource>,
@@ -122,20 +120,8 @@ pub struct FontSources {
     masks: Vec<BitSet>,
 }
 
-impl FontSources {
-    /// Discovers and loads fonts from the specified directories by recursively scanning for `.ttf`, `.otf`, and `.ttc` files.
-    pub fn resolve(config: &mut OptOneMany<PathBuf>) -> Result<Self, FontError> {
-        if config.is_empty() {
-            return Ok(Self::default());
-        }
-
-        let mut fonts = DashMap::new();
-        let lib = Library::init()?;
-
-        for path in config.iter() {
-            recurse_dirs(&lib, path.clone(), &mut fonts, true)?;
-        }
-
+impl Default for FontSources {
+    fn default() -> Self {
         let mut masks = Vec::with_capacity(MAX_UNICODE_CP_RANGE_ID + 1);
 
         let mut bs = BitSet::with_capacity(CP_RANGE_SIZE);
@@ -147,7 +133,18 @@ impl FontSources {
             }
         }
 
-        Ok(Self { fonts, masks })
+        Self {
+            fonts: DashMap::new(),
+            masks,
+        }
+    }
+}
+
+impl FontSources {
+    /// Discovers and loads fonts from the specified directory by recursively scanning for `.ttf`, `.otf`, and `.ttc` files.
+    pub fn recursively_add_directory(&mut self, path: PathBuf) -> Result<(), FontError> {
+        let lib = Library::init()?;
+        recurse_dirs(&lib, path, &mut self.fonts, true)
     }
 
     /// Returns a catalog of all loaded fonts
