+++ martin/src/cog/mod.rs
@@ -1,10 +1,9 @@
 #![doc = include_str!("README.md")]
 
-mod config;
 mod errors;
 mod image;
 mod model;
 mod source;
 
-pub use config::CogConfig;
 pub use errors::CogError;
+pub use source::CogSource;
