+++ martin/src/config/args/connections.rs
@@ -22,11 +22,23 @@ impl Arguments {
         }
     }
 
+    #[cfg(any(
+        feature = "postgres",
+        feature = "pmtiles",
+        feature = "mbtiles",
+        feature = "cog"
+    ))]
     #[must_use]
     pub fn is_empty(&self) -> bool {
         self.values.is_empty()
     }
 
+    #[cfg(any(
+        feature = "postgres",
+        feature = "pmtiles",
+        feature = "mbtiles",
+        feature = "cog"
+    ))]
     /// Process input params in the original order, but only if no prior consumer has claimed it.
     /// Each consumer can either take it (no other consumer will see it),
     /// share it (other consumers will see it too),
