+++ martin/src/config/file/pg/config_function.rs
@@ -1,8 +1,8 @@
 use serde::{Deserialize, Serialize};
 use tilejson::{Bounds, TileJSON};
 
-use crate::config::UnrecognizedValues;
-use crate::pg::config::PgInfo;
+use super::config::PgInfo;
+use crate::config::file::UnrecognizedValues;
 use crate::pg::utils::{InfoMap, patch_json};
 
 pub type FuncInfoSources = InfoMap<FunctionInfo>;
