+++ martin/src/config/file/pg/mod.rs
@@ -1,0 +1,8 @@
+mod config;
+pub use config::*;
+
+mod config_function;
+pub use config_function::*;
+
+mod config_table;
+pub use config_table::*;
