+++ martin/src/lib.rs
@@ -1,8 +1,7 @@
 #![cfg_attr(doc, doc = include_str!("../README.md"))]
 #![forbid(unsafe_code)]
 
-mod config;
-pub use config::{Config, ServerState, read_config};
+pub mod config;
 
 mod source;
 pub use source::{Source, TileData, TileInfoSource, TileSources, UrlQuery};
@@ -10,10 +9,8 @@ pub use source::{Source, TileData, TileInfoSource, TileSources, UrlQuery};
 mod utils;
 pub use utils::{IdResolver, MartinError, MartinResult, NO_MAIN_CACHE};
 
-pub mod args;
 #[cfg(feature = "cog")]
 pub mod cog;
-pub mod file_config;
 #[cfg(feature = "mbtiles")]
 pub mod mbtiles;
 #[cfg(feature = "postgres")]
