+++ martin/src/pg/query_functions.rs
@@ -5,10 +5,10 @@ use log::{debug, warn};
 use postgres_protocol::escape::escape_identifier;
 use serde_json::Value;
 
+use crate::config::file::pg::FunctionInfo;
 use crate::pg::PgError::PostgresError;
 use crate::pg::PgResult;
 use crate::pg::builder::SqlFuncInfoMapMap;
-use crate::pg::config_function::FunctionInfo;
 use crate::pg::pg_source::PgSqlInfo;
 use crate::pg::pool::PgPool;
