+++ martin/src/pg/query_tables.rs
@@ -9,12 +9,11 @@ use serde_json::Value;
 use tilejson::Bounds;
 use tokio::time::timeout;
 
-use crate::args::{BoundsCalcType, DEFAULT_BOUNDS_TIMEOUT};
+use crate::config::args::{BoundsCalcType, DEFAULT_BOUNDS_TIMEOUT};
+use crate::config::file::pg::{PgInfo, TableInfo};
 use crate::pg::PgError::PostgresError;
 use crate::pg::PgResult;
 use crate::pg::builder::SqlTableInfoMapMapMap;
-use crate::pg::config::PgInfo;
-use crate::pg::config_table::TableInfo;
 use crate::pg::pg_source::PgSqlInfo;
 use crate::pg::pool::PgPool;
 use crate::pg::utils::{json_to_hashmap, polygon_to_bbox};
