+++ martin/src/pg/tls.rs
@@ -17,11 +17,12 @@ use rustls_native_certs::load_native_certs;
 use rustls_pemfile::Item::Pkcs1Key;
 use tokio_postgres_rustls::MakeRustlsConnect;
 
+use crate::config::file::pg::PgSslCerts;
 use crate::pg::PgError::{
     BadConnectionString, CannotLoadRoots, CannotOpenCert, CannotParseCert, CannotUseClientKey,
     InvalidPrivateKey, UnknownSslMode,
 };
-use crate::pg::{PgResult, PgSslCerts};
+use crate::pg::PgResult;
 
 /// A temporary workaround for <https://github.com/sfackler/rust-postgres/pull/988>
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
