+++ martin/src/pg/utils.rs
@@ -1,14 +1,10 @@
 use std::collections::{BTreeMap, HashMap};
-use std::future::Future;
-use std::time::Duration;
 
 use deadpool_postgres::tokio_postgres::types::Json;
-use futures::pin_mut;
 use itertools::Itertools as _;
 use log::{error, info, warn};
 use postgis::{LineString, Point, Polygon, ewkb};
 use tilejson::{Bounds, TileJSON};
-use tokio::time::timeout;
 
 use crate::source::UrlQuery;
 
@@ -28,20 +24,6 @@ pub fn sorted_opt_set<S: serde::Serializer>(
             v
         })
         .serialize(serializer)
-}
-
-pub async fn on_slow<T, S: FnOnce()>(
-    future: impl Future<Output = T>,
-    duration: Duration,
-    fn_on_slow: S,
-) -> T {
-    pin_mut!(future);
-    if let Ok(result) = timeout(duration, &mut future).await {
-        result
-    } else {
-        fn_on_slow();
-        future.await
-    }
 }
 
 #[must_use]
