+++ martin/src/srv/mod.rs
@@ -1,13 +1,8 @@
-mod config;
-pub use config::{KEEP_ALIVE_DEFAULT, LISTEN_ADDRESSES_DEFAULT, SrvConfig};
-
 #[cfg(feature = "fonts")]
 mod fonts;
 
 mod server;
 pub use server::{Catalog, RESERVED_KEYWORDS, new_server, router};
-
-pub mod cors;
 
 mod tiles;
 pub use tiles::{DynTileSource, TileRequest};
