+++ martin/src/utils/error.rs
@@ -71,7 +71,7 @@ pub enum MartinError {
     CogError(#[from] crate::cog::CogError),
 
     #[error(transparent)]
-    FileError(#[from] crate::file_config::FileError),
+    ConfigFileError(#[from] crate::config::file::ConfigFileError),
 
     #[cfg(feature = "sprites")]
     #[error(transparent)]
@@ -93,7 +93,4 @@ pub enum MartinError {
 
     #[error("Internal error: {0}")]
     InternalError(#[from] Box<dyn Error + Send + Sync>),
-
-    #[error(transparent)]
-    CorsError(#[from] crate::srv::cors::CorsError),
 }
