+++ martin/tests/utils/mod.rs
@@ -3,11 +3,10 @@
 #![allow(clippy::unused_async)]
 
 mod pg_utils;
-
 use actix_web::dev::ServiceResponse;
 use actix_web::test::read_body;
 use log::warn;
-use martin::Config;
+use martin::config::file::Config;
 use martin_core::config::env::FauxEnv;
 pub use pg_utils::*;
